Refine retry option validation
